Remove keychain
Do not use keychain on untrusted systems. Keychain starts the ssh-agent
and gpg-agent automatically and caches both these types of keys. This is
not a problem for ssh because it cooperates with ssh key forwarding
which still avoids having private keys files on untrusted systems. But
having a gpg-agent started prevents the gpg agent forwarding. Of course,
by not starting either of the agents we also avoid storing any keys in
memory on these systems too.